webpack complete

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
-import Video, { formatHashtags } from "../models/Video";
+import Video, { formatHashtags } from "../models/Video.js";
+import User from "../models/User.js";
 
 export const home = async (req, res) => {
   try {
@@ -10,28 +11,47 @@ export const home = async (req, res) => {
 };
 export const watch = async (req, res) => {
   const { id } = req.params;
+  const video = await Video.findById(id).populate("owner");
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+
+  return res.render("watch", {
+    pageTitle: video.title,
+    video,
+  });
+};
+export const getEdit = async (req, res) => {
+  const {
+    params: { id },
+    session: {
+      user: { _id },
+    },
+  } = req;
+
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
   }
-  return res.render("watch", { pageTitle: `Watching ${video.title}`, video });
-};
-export const getEdit = async (req, res) => {
-  const { id } = req.params;
-  const video = await Video.findById(id);
-  if (!video) {
-    return res.status(404).render("404", { pageTitle: "Video not found" });
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   return res.render("edit", { pageTitle: `Editing ${video.title}`, video });
 };
 export const postEdit = async (req, res) => {
-  const { id } = req.params;
   const {
+    params: { id },
+    session: {
+      user: { _id },
+    },
     body: { title, description, hashtags },
   } = req;
   const video = await Video.exists({ _id: id });
   if (!video) {
     return res.render("404", { pageTitle: "Video not found" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -44,13 +64,25 @@ export const getUpload = (req, res) => {
   res.render("upload", { pageTitle: "Upload" });
 };
 export const postUpload = async (req, res) => {
-  const { title, description, hashtags } = req.body;
+  const {
+    file: { path },
+    session: {
+      user: { _id },
+    },
+    body: { title, description, hashtags },
+  } = req;
+
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
+      fileUrl: path,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     console.log(error);
@@ -61,7 +93,19 @@ export const postUpload = async (req, res) => {
   }
 };
 export const deleteVideo = async (req, res) => {
-  const { id } = req.params;
+  const {
+    params: { id },
+    session: {
+      user: { _id },
+    },
+  } = req;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.render("404", { pageTitle: "Video not found" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
